requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.2.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.66.3
packaging==25.0
pika==1.3.2
psutil==7.0.0
psycopg2-binary==2.9.9
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
win32_setctime==1.2.0
